Refactor Tool.java into different classes per cyin suggestion
