+++ tool/pom.xml
@@ -66,7 +66,7 @@
                     <archive>
                         <manifest>
                             <addClasspath>true</addClasspath>
-                            <mainClass>org.datacommons.tool.Tool</mainClass>
+                            <mainClass>org.datacommons.tool.Main</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
