+++ tool/src/main/java/org/datacommons/tool/Lint.java
@@ -1,0 +1,89 @@
+package org.datacommons.tool;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.Callable;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.datacommons.proto.Debug;
+import org.datacommons.proto.Mcf;
+import picocli.CommandLine;
+
+@CommandLine.Command(name = "lint", description = "Run various checks on input MCF/TMCF/CSV files")
+class Lint implements Callable<Integer> {
+  private static final Logger logger = LogManager.getLogger(Lint.class);
+
+  @CommandLine.Parameters(
+      arity = "1..*",
+      description =
+          ("List of input files. The file extensions are used to infer the format. "
+              + "Valid extensions include .mcf for Instance MCF, .tmcf for Template MCF, "
+              + ".csv for tabular text files delimited by comma (overridden with -d), and .tsv "
+              + "for tab-delimited tabular files."))
+  private File[] files;
+
+  @CommandLine.Option(
+      names = {"-d", "--delimiter"},
+      description =
+          "Delimiter of the input CSV files. Default is ',' for .csv files and '\\t' for "
+              + ".tsv files.",
+      scope = CommandLine.ScopeType.INHERIT)
+  private Character delimiter;
+
+  @CommandLine.Spec CommandLine.Model.CommandSpec spec; // injected by picocli
+
+  @Override
+  public Integer call() throws IOException {
+    List<File> mcfFiles = new ArrayList<>();
+    List<File> tmcfFiles = new ArrayList<>();
+    List<File> csvFiles = new ArrayList<>();
+    int nTsv = 0;
+    for (File file : files) {
+      String lowerPath = file.getPath().toLowerCase();
+      if (lowerPath.endsWith(".mcf")) {
+        mcfFiles.add(file);
+      } else if (lowerPath.endsWith(".tmcf")) {
+        tmcfFiles.add(file);
+      } else if (lowerPath.endsWith(".csv")) {
+        csvFiles.add(file);
+      } else if (lowerPath.endsWith(".tsv")) {
+        nTsv++;
+        csvFiles.add(file);
+      } else {
+        throw new CommandLine.ParameterException(
+            spec.commandLine(), "Found an unsupported file type: " + file.getPath());
+      }
+    }
+    logger.info(
+        "Input includes {} MCF file(s), {} TMCF file(s), {} CSV file(s)",
+        mcfFiles.size(),
+        tmcfFiles.size(),
+        csvFiles.size());
+    // Various checks
+    if (nTsv > 0 && nTsv != csvFiles.size()) {
+      throw new CommandLine.ParameterException(
+          spec.commandLine(), "Please do not mix .tsv and .csv files");
+    }
+    if (delimiter == null) {
+      delimiter = nTsv > 0 ? '\t' : ',';
+    }
+    if (!csvFiles.isEmpty() && tmcfFiles.size() != 1) {
+      throw new CommandLine.ParameterException(
+          spec.commandLine(), "Please provide one .tmcf file with CSV/TSV files");
+    }
+    Debug.Log.Builder logCtx = Debug.Log.newBuilder();
+    for (File f : mcfFiles) {
+      Processor.processNodes(Mcf.McfType.INSTANCE_MCF, f, logCtx, logger);
+    }
+    if (!csvFiles.isEmpty()) {
+      Processor.processTables(tmcfFiles.get(0), csvFiles, delimiter, null, logCtx, logger);
+    } else {
+      for (File f : tmcfFiles) {
+        Processor.processNodes(Mcf.McfType.TEMPLATE_MCF, f, logCtx, logger);
+      }
+    }
+    return 0;
+  }
+}
