+++ tool/src/main/java/org/datacommons/tool/Processor.java
@@ -1,0 +1,57 @@
+package org.datacommons.tool;
+
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.IOException;
+import java.util.List;
+import org.apache.logging.log4j.Logger;
+import org.datacommons.proto.Debug;
+import org.datacommons.proto.Mcf;
+import org.datacommons.util.McfParser;
+import org.datacommons.util.McfUtil;
+import org.datacommons.util.TmcfCsvParser;
+
+public class Processor {
+  public static void processNodes(
+      Mcf.McfType type, File file, Debug.Log.Builder logCtx, Logger logger) throws IOException {
+    int numNodesProcessed = 0;
+    logger.debug("Checking {}", file.getName());
+    // TODO: isResolved is more allowing, be stricter.
+    McfParser parser = McfParser.init(type, file.getPath(), false);
+    Mcf.McfGraph n;
+    while ((n = parser.parseNextNode()) != null) {
+      numNodesProcessed++;
+    }
+    logger.info("Checked {} with {} nodes", file.getName(), numNodesProcessed);
+  }
+
+  public static void processTables(
+      File tmcfFile,
+      List<File> csvFiles,
+      char delimiter,
+      BufferedWriter writer,
+      Debug.Log.Builder logCtx,
+      Logger logger)
+      throws IOException {
+    logger.debug("TMCF " + tmcfFile.getName());
+    for (File csvFile : csvFiles) {
+      logger.debug("Checking CSV " + csvFile.getPath());
+      TmcfCsvParser parser =
+          TmcfCsvParser.init(tmcfFile.getPath(), csvFile.getPath(), delimiter, logCtx);
+      Mcf.McfGraph g;
+      int numNodesProcessed = 0, numRowsProcessed = 0;
+      while ((g = parser.parseNextRow()) != null) {
+        numNodesProcessed += g.getNodesCount();
+        numRowsProcessed++;
+        if (writer != null) {
+          writer.write(McfUtil.serializeMcfGraph(g, false));
+        }
+      }
+      logger.info(
+          "Checked CSV {} ({} rows, {} nodes)",
+          csvFile.getName(),
+          numRowsProcessed,
+          numNodesProcessed);
+    }
+  }
+}
